requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.9.1
PyMySQL==1.1.0
python-dotenv==1.0.17